Starting changes to allow for multiple scans and comparing against multiple scans
